test passing fixes

+++ src/test/java/NumberUtilTest.java
@@ -7,11 +7,11 @@ public class NumberUtilTest {
 
     @Test
     public void maxTest() {
-        assertEquals(max(1, 5, 10), 10);
+        assertEquals(max(1, 5, 10).intValue(), 10);
     }
 
     @Test
     public void minTest() {
-        assertEquals(min(1, 5, 10), 1);
+        assertEquals(min(1, 5, 10).intValue(), 1);
     }
 }
